+++ resources/openjdk_source_install.rb
@@ -1,15 +1,11 @@
 provides :openjdk_source_install
 unified_mode true
 include Java::Cookbook::OpenJdkHelpers
+include Java::Cookbook::BinCmdHelpers
 
 property :version, String,
           name_property: true,
           description: 'Java version to install'
-
-property :variant, String,
-          equal_to: %w(openjdk semeru temurin),
-          default: 'openjdk',
-          description: 'Install flavour'
 
 property :url, String,
           default: lazy { default_openjdk_url(version, variant) },
@@ -25,12 +21,13 @@ property :java_home, String,
           description: 'Set to override the java_home'
 
 property :bin_cmds, Array,
-          default: lazy { default_openjdk_bin_cmds(version) },
+          default: lazy { default_bin_cmds(version) },
           description: 'A list of bin_cmds based on the version and variant'
 
 use 'partial/_common'
 use 'partial/_linux'
 use 'partial/_java_home'
+use 'partial/_openjdk'
 
 action :install do
   extract_dir = new_resource.java_home.split('/')[0..-2].join('/')