+++ CHANGELOG.md
@@ -4,6 +4,17 @@ This file is used to list changes made in each version of the Java cookbook.
 
 ## Unreleased
 
+- Add new resource `temurin_package_install`
+- Add script to check for Java updates
+- Update Temurin Java 8 support
+- Update Temurin repositories
+- Update bin commands for all OpenJDK versions
+- Fix Java alternatives to prevent unnecessary removal and re-addition of alternatives
+- Move bin_cmds from Java::Cookbook::OpenJdkHelpers to Java::Cookbook::BinCmdHelpers for reuse outside of OpenJDK
+- Fix apt_repository failing to install the GPG in the correct location
+- Add Temurin 21 to the test matrix
+- Remove Semeru from the test matrix
+
 ## 12.1.1 - *2024-12-05*
 
 ## 12.1.0 - *2024-12-03*